# Cleanup artifacts before upload