--- a/dev/tools/list-potential-artifacts.sh
+++ b/dev/tools/list-potential-artifacts.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+
+for d in _build_ci stdlib/_build; do
+  if [ -d $d ]; then
+    find $d -type f -o -type l | sort
+  fi
+done