We remove the files which were already present when the job started.