@SkySkimmer 2024-12-13
Seems to work, HB failure is unrelated
example master pipeline https://gitlab.inria.fr/coq/coq/-/pipelines/1085613
PR pipeline https://gitlab.inria.fr/coq/coq/-/pipelines/1085818
~~~
|             | master size  | master time | PR size      | PR time   |
| total       | didn't check | 130min      | didn't check | 93min     |
| aac_tactics | 51.3MB       | 1min 48s    | 3.0MB        | 1min 10s  |
| metacoq     | 235MB        | 36min 11s   | 177MB        | 29min 20s |
| fiat_crypto | 366MB        | 41min 5s    | 172MB        | 32min 40s |
~~~
(NB there is significant variance between timings and 90min isn't unknown in master, eg https://gitlab.inria.fr/coq/coq/-/pipelines/1084787, but at least we shouldn't be slower)