@SkySkimmer 2024-12-13
@JasonGross @Zimmi48 not sure if the minimizer needs to be changed for this