@SkySkimmer 2024-12-13
More stats:
https://gitlab.inria.fr/coq/coq/-/pipelines/1084116 (master before #19921):
~~~
| name                             | duration   | artifacts size   |      id |
|----------------------------------+------------+------------------+---------|
| library:ci-coqtail               | 45s        | 167.33 KiB       | 5061166 |
| validate:base+32bit              | 52s        | 13.79 KiB        | 5061159 |
| plugin:ci-reduction_effects      | 56s        | 159.71 KiB       | 5061200 |
| validate:edge+flambda            | 58s        | 13.79 KiB        | 5061160 |
| plugin:ci-tactician              | 59s        | 6.11 MiB         | 5061206 |
| validate:base                    | 1min 2s    | 13.71 KiB        | 5061158 |
| plugin:ci-ltac2_compiler         | 1min 2s    | 731.58 KiB       | 5061196 |
| plugin:plugin-tutorial           | 1min 17s   | 0.00 B           | 5061148 |
| doc:init                         | 1min 21s   | 1.73 MiB         | 5061151 |
| doc:ml-api:odoc                  | 1min 21s   | 5.83 MiB         | 5061152 |
| plugin:ci-lean_importer          | 1min 23s   | 15.99 MiB        | 5061195 |
| plugin:ci-vscoq                  | 2min 44s   | 170.20 MiB       | 5061203 |
| lint                             | 2min 8s    | 0.00 B           | 5061145 |
| plugin:ci-autosubst_ocaml        | 2min 18s   | 38.76 MiB        | 5061186 |
| pkg:nix                          | 3min 2s    | 0.00 B           | 5061147 |
| build:base+32bit                 | 3min 17s   | 213.15 MiB       | 5061140 |
| build:base                       | 3min 22s   | 222.66 MiB       | 5061139 |
| build:base+async                 | 4min 39s   | 222.66 MiB       | 5061144 |
| doc:refman                       | 4min 5s    | 12.38 MiB        | 5061149 |
| build:base:dev                   | 4min 9s    | 235.29 MiB       | 5061142 |
| library:ci-stdlib                | 4min 12s   | 0.00 B           | 5061179 |
| library:ci-hott                  | 5min 53s   | 22.28 MiB        | 5061175 |
| library:ci-itree_io              | 5min 60s   | 42.27 MiB        | 5061240 |
| plugin:ci-equations              | 6min 21s   | 9.76 MiB         | 5061193 |
| plugin:ci-paramcoq               | 6min 28s   | 3.96 MiB         | 5061198 |
| library:ci-stdlib+flambda        | 7min 35s   | 0.00 B           | 5061180 |
| plugin:ci-coqhammer              | 7min 36s   | 6.65 MiB         | 5061191 |
| library:ci-mathcomp_word         | 7min 44s   | 169.73 MiB       | 5061220 |
| test-suite:base                  | 7min 47s   | 0.00 B           | 5061153 |
| plugin:ci-coq_dpdgraph           | 7min 47s   | 4.14 MiB         | 5061190 |
| library:ci-autosubst             | 7min 52s   | 626.49 KiB       | 5061162 |
| library:ci-bbv                   | 7min 57s   | 1.54 MiB         | 5061163 |
| plugin:ci-coinduction            | 7min 3s    | 990.05 KiB       | 5061189 |
| library:ci-simple_io             | 7min 3s    | 7.79 MiB         | 5061217 |
| doc:refman-pdf                   | 7min 4s    | 11.75 MiB        | 5061150 |
| plugin:ci-mtac2                  | 7min 9s    | 13.83 MiB        | 5061197 |
| library:ci-argosy                | 7min 10s   | 2.99 MiB         | 5061161 |
| library:ci-trakt                 | 7min 16s   | 84.85 MiB        | 5061221 |
| library:ci-coquelicot            | 8min 38s   | 176.17 MiB       | 5061229 |
| test-suite:base+async            | 8min 50s   | 0.00 B           | 5061157 |
| library:ci-fcsl_pcm              | 8min 52s   | 181.40 MiB       | 5061212 |
| library:ci-ext_lib               | 8min 56s   | 6.70 MiB         | 5061170 |
| plugin:ci-stalmarck              | 8min 57s   | 5.50 MiB         | 5061205 |
| plugin:ci-elpi_hb                | 8min 60s   | 84.30 MiB        | 5061192 |
| build:base:dev:dune              | 8min 0s    | 0.00 B           | 5061143 |
| plugin:ci-waterproof             | 8min 2s    | 3.19 MiB         | 5061207 |
| plugin:ci-aac_tactics            | 8min 10s   | 3.01 MiB         | 5061184 |
| library:ci-stdlib_doc            | 8min 10s   | 0.00 B           | 5061219 |
| plugin:ci-hb_test                | 8min 11s   | 97.02 MiB        | 5061223 |
| plugin:ci-elpi_test              | 8min 15s   | 114.29 MiB       | 5061222 |
| library:ci-menhir                | 8min 22s   | 56.74 MiB        | 5061174 |
| test-suite:base:dev              | 8min 23s   | 0.00 B           | 5061156 |
| library:ci-paco                  | 8min 26s   | 8.11 MiB         | 5061177 |
| plugin:ci-bignums                | 8min 26s   | 7.51 MiB         | 5061188 |
| plugin:ci-smtcoq                 | 8min 28s   | 13.94 MiB        | 5061204 |
| plugin:ci-equations_test         | 9min 34s   | 18.13 MiB        | 5061224 |
| library:ci-tlc                   | 9min 40s   | 6.15 MiB         | 5061181 |
| build:edge+flambda               | 9min 44s   | 396.53 MiB       | 5061141 |
| library:ci-neural_net_interp     | 9min 45s   | 24.63 MiB        | 5061176 |
| plugin:ci-coq_lsp                | 9min 2s    | 202.82 MiB       | 5061202 |
| library:ci-coqutil               | 9min 7s    | 17.58 MiB        | 5061167 |
| plugin:ci-smtcoq_trakt           | 9min 15s   | 103.53 MiB       | 5061249 |
| library:ci-coqprime              | 9min 17s   | 17.21 MiB        | 5061211 |
| library:ci-flocq                 | 9min 19s   | 8.89 MiB         | 5061172 |
| library:ci-math_classes          | 9min 25s   | 20.00 MiB        | 5061214 |
| plugin:ci-atbr                   | 9min 29s   | 12.74 MiB        | 5061185 |
| test-suite:base+32bit            | 10min 35s  | 0.00 B           | 5061154 |
| library:ci-stdlib_test           | 10min 42s  | 0.00 B           | 5061218 |
| library:ci-sf                    | 10min 47s  | 85.16 MiB        | 5061178 |
| library:ci-http                  | 10min 53s  | 286.76 MiB       | 5061244 |
| doc:refman:deploy                | 10min 56s  | 0.00 B           | 5061251 |
| plugin:ci-itauto                 | 10min 6s   | 12.27 MiB        | 5061187 |
| library:ci-engine_bench          | 11min 31s  | 2.17 MiB         | 5061169 |
| library:ci-kami                  | 11min 52s  | 27.53 MiB        | 5061173 |
| test-suite:edge+flambda          | 12min 49s  | 0.00 B           | 5061155 |
| library:ci-color                 | 12min 56s  | 67.17 MiB        | 5061209 |
| plugin:ci-rewriter               | 13min 52s  | 16.30 MiB        | 5061201 |
| library:ci-riscv_coq             | 13min 12s  | 41.15 MiB        | 5061226 |
| library:ci-cross_crypto          | 14min 34s  | 13.04 MiB        | 5061168 |
| library:ci-category_theory       | 14min 38s  | 21.40 MiB        | 5061208 |
| plugin:ci-quickchick_test        | 14min 3s   | 196.62 MiB       | 5061246 |
| library:ci-rupicola              | 14min 24s  | 120.56 MiB       | 5061248 |
| plugin:ci-fiat_parsers           | 15min 57s  | 16.57 MiB        | 5061194 |
| library:ci-bigenough             | 15min 9s   | 169.23 MiB       | 5061238 |
| library:ci-mczify                | 15min 18s  | 171.48 MiB       | 5061236 |
| library:ci-finmap                | 15min 20s  | 170.98 MiB       | 5061237 |
| library:ci-deriving              | 16min 30s  | 172.90 MiB       | 5061242 |
| library:ci-compcert              | 16min 10s  | 148.95 MiB       | 5061210 |
| plugin:ci-quickchick             | 16min 28s  | 190.75 MiB       | 5061245 |
| library:ci-jasmin                | 17min 39s  | 206.65 MiB       | 5061243 |
| library:ci-corn                  | 17min 7s   | 123.06 MiB       | 5061234 |
| plugin:ci-relation_algebra       | 17min 10s  | 180.58 MiB       | 5061247 |
| library:ci-mathcomp_test         | 17min 14s  | 169.36 MiB       | 5061235 |
| library:ci-oddorder              | 17min 22s  | 189.49 MiB       | 5061232 |
| library:ci-itree                 | 17min 27s  | 41.28 MiB        | 5061216 |
| library:ci-verdi_raft            | 17min 28s  | 36.48 MiB        | 5061183 |
| pkg:opam:native                  | 18min 33s  | 19.57 KiB        | 5061146 |
| library:ci-bedrock2              | 18min 3s   | 102.50 MiB       | 5061227 |
| library:ci-mathcomp              | 20min 42s  | 169.16 MiB       | 5061215 |
| library:ci-coq_tools             | 20min 49s  | 2.79 MiB         | 5061165 |
| library:ci-fiat_crypto_ocaml     | 21min 54s  | 0.00 B           | 5061231 |
| library:ci-coq_performance_tests | 24min 11s  | 11.18 MiB        | 5061164 |
| library:ci-iris                  | 25min 27s  | 59.70 MiB        | 5061213 |
| library:ci-fourcolor             | 29min 25s  | 188.10 MiB       | 5061233 |
| plugin:ci-perennial              | 30min 56s  | 64.22 MiB        | 5061199 |
| library:ci-analysis              | 32min 29s  | 217.33 MiB       | 5061239 |
| library:ci-fiat_crypto_legacy    | 37min 51s  | 58.28 MiB        | 5061171 |
| library:ci-unimath               | 38min 47s  | 45.79 MiB        | 5061182 |
| plugin:ci-metacoq                | 38min 51s  | 186.72 MiB       | 5061225 |
| doc:ci-refman                    | 43min 54s  | 24.68 MiB        | 5061250 |
| library:ci-vst                   | 53min 0s   | 227.61 MiB       | 5061241 |
| library:ci-fiat_crypto           | 57min 31s  | 318.62 MiB       | 5061230 |
| library:ci-bedrock2_examples     | 1h 4min    | 138.31 MiB       | 5061228 |
| total                            | 23h 55min  | 7.62 GiB         | 1084116 |
~~~

current master https://gitlab.inria.fr/coq/coq/-/pipelines/1085613
~~~
| name                             | duration   | artifacts size   |      id |
|----------------------------------+------------+------------------+---------|
| validate:base+32bit              | 54s        | 13.79 KiB        | 5072720 |
| plugin:ci-ltac2_compiler         | 56s        | 731.55 KiB       | 5072757 |
| library:ci-coqtail               | 1min 1s    | 167.33 KiB       | 5072727 |
| plugin:ci-reduction_effects      | 1min 2s    | 159.71 KiB       | 5072761 |
| plugin:ci-lean_importer          | 1min 14s   | 15.99 MiB        | 5072756 |
| plugin:ci-paramcoq               | 1min 17s   | 52.34 MiB        | 5072759 |
| plugin:ci-coq_dpdgraph           | 1min 26s   | 52.52 MiB        | 5072751 |
| plugin:ci-coinduction            | 1min 27s   | 49.35 MiB        | 5072750 |
| library:ci-menhir                | 2min 31s   | 105.12 MiB       | 5072735 |
| validate:base                    | 2min 39s   | 13.71 KiB        | 5072719 |
| plugin:plugin-tutorial           | 2min 39s   | 0.00 B           | 5072709 |
| library:ci-stdlib_doc            | 2min 42s   | 52.53 MiB        | 5072780 |
| library:ci-trakt                 | 2min 46s   | 133.23 MiB       | 5072782 |
| plugin:ci-aac_tactics            | 2min 49s   | 51.39 MiB        | 5072745 |
| plugin:ci-waterproof             | 2min 49s   | 51.58 MiB        | 5072768 |
| library:ci-tlc                   | 2min 50s   | 50.39 MiB        | 5072742 |
| plugin:ci-bignums                | 2min 55s   | 55.90 MiB        | 5072749 |
| plugin:ci-vscoq                  | 2min 58s   | 170.34 MiB       | 5072764 |
| library:ci-autosubst             | 2min 4s    | 49.00 MiB        | 5072723 |
| library:ci-argosy                | 2min 9s    | 47.23 MiB        | 5072722 |
| lint                             | 2min 13s   | 0.00 B           | 5072706 |
| plugin:ci-stalmarck              | 2min 15s   | 49.74 MiB        | 5072766 |
| doc:init                         | 2min 19s   | 1.73 MiB         | 5072712 |
| plugin:ci-equations_test         | 2min 26s   | 66.51 MiB        | 5072785 |
| library:ci-bigenough             | 2min 28s   | 217.59 MiB       | 5072799 |
| build:base+32bit                 | 2min 28s   | 213.13 MiB       | 5072701 |
| plugin:ci-mtac2                  | 2min 29s   | 62.22 MiB        | 5072758 |
| library:ci-bbv                   | 3min 47s   | 45.78 MiB        | 5072724 |
| build:base                       | 3min 53s   | 222.63 MiB       | 5072700 |
| validate:edge+flambda            | 3min 54s   | 13.71 KiB        | 5072721 |
| plugin:ci-quickchick             | 3min 56s   | 239.12 MiB       | 5072806 |
| plugin:ci-atbr                   | 3min 57s   | 61.12 MiB        | 5072746 |
| library:ci-coqutil               | 3min 58s   | 65.96 MiB        | 5072728 |
| plugin:ci-tactician              | 3min 0s    | 6.11 MiB         | 5072767 |
| doc:ml-api:odoc                  | 3min 10s   | 5.83 MiB         | 5072713 |
| plugin:ci-autosubst_ocaml        | 3min 13s   | 100.03 MiB       | 5072747 |
| library:ci-itree_io              | 3min 15s   | 90.66 MiB        | 5072801 |
| plugin:ci-smtcoq_trakt           | 3min 15s   | 151.90 MiB       | 5072810 |
| library:ci-mathcomp_test         | 3min 18s   | 217.72 MiB       | 5072796 |
| library:ci-neural_net_interp     | 3min 19s   | 68.87 MiB        | 5072737 |
| plugin:ci-coqhammer              | 3min 20s   | 55.04 MiB        | 5072752 |
| library:ci-deriving              | 4min 33s   | 221.27 MiB       | 5072803 |
| plugin:ci-elpi_hb                | 4min 38s   | 132.67 MiB       | 5072753 |
| library:ci-simple_io             | 4min 38s   | 56.17 MiB        | 5072778 |
| library:ci-ext_lib               | 4min 38s   | 55.08 MiB        | 5072731 |
| plugin:ci-smtcoq                 | 4min 42s   | 58.18 MiB        | 5072765 |
| plugin:ci-quickchick_test        | 4min 43s   | 244.98 MiB       | 5072807 |
| library:ci-flocq                 | 4min 46s   | 57.28 MiB        | 5072733 |
| plugin:ci-relation_algebra       | 4min 51s   | 228.94 MiB       | 5072808 |
| library:ci-paco                  | 4min 0s    | 56.49 MiB        | 5072738 |
| build:base+async                 | 4min 3s    | 222.63 MiB       | 5072705 |
| library:ci-http                  | 4min 6s    | 335.14 MiB       | 5072805 |
| plugin:ci-itauto                 | 4min 7s    | 60.65 MiB        | 5072748 |
| build:base:dev:dune              | 4min 9s    | 0.00 B           | 5072704 |
| library:ci-sf                    | 4min 17s   | 129.40 MiB       | 5072739 |
| pkg:nix                          | 4min 29s   | 0.00 B           | 5072708 |
| plugin:ci-equations              | 5min 51s   | 58.14 MiB        | 5072754 |
| library:ci-hott                  | 5min 56s   | 22.42 MiB        | 5072736 |
| build:base:dev                   | 5min 57s   | 234.98 MiB       | 5072703 |
| library:ci-coqprime              | 5min 4s    | 65.60 MiB        | 5072772 |
| library:ci-color                 | 5min 6s    | 115.56 MiB       | 5072770 |
| plugin:ci-hb_test                | 5min 13s   | 145.79 MiB       | 5072784 |
| library:ci-math_classes          | 5min 28s   | 68.38 MiB        | 5072775 |
| doc:refman                       | 5min 29s   | 12.38 MiB        | 5072710 |
| library:ci-kami                  | 5min 29s   | 75.92 MiB        | 5072734 |
| library:ci-engine_bench          | 6min 32s   | 46.41 MiB        | 5072730 |
| library:ci-rupicola              | 6min 37s   | 168.95 MiB       | 5072809 |
| library:ci-mathcomp_word         | 6min 38s   | 218.09 MiB       | 5072781 |
| plugin:ci-elpi_test              | 6min 45s   | 162.66 MiB       | 5072783 |
| library:ci-finmap                | 6min 49s   | 219.35 MiB       | 5072798 |
| plugin:ci-coq_lsp                | 6min 52s   | 251.16 MiB       | 5072763 |
| doc:refman:deploy                | 6min 54s   | 0.00 B           | 5072812 |
| library:ci-fcsl_pcm              | 6min 1s    | 229.76 MiB       | 5072773 |
| library:ci-coquelicot            | 6min 14s   | 224.53 MiB       | 5072790 |
| library:ci-stdlib+flambda        | 6min 19s   | 61.29 MiB        | 5072741 |
| library:ci-cross_crypto          | 6min 26s   | 57.28 MiB        | 5072729 |
| library:ci-stdlib                | 7min 35s   | 53.27 MiB        | 5072740 |
| library:ci-mczify                | 7min 35s   | 219.85 MiB       | 5072797 |
| test-suite:base:dev              | 7min 39s   | 0.00 B           | 5072717 |
| doc:refman-pdf                   | 7min 49s   | 11.74 MiB        | 5072711 |
| library:ci-stdlib_test           | 7min 8s    | 44.24 MiB        | 5072779 |
| plugin:ci-rewriter               | 8min 48s   | 64.69 MiB        | 5072762 |
| test-suite:base                  | 9min 35s   | 0.00 B           | 5072714 |
| library:ci-compcert              | 9min 37s   | 197.33 MiB       | 5072771 |
| build:edge+flambda               | 9min 42s   | 396.43 MiB       | 5072702 |
| test-suite:base+32bit            | 9min 12s   | 0.00 B           | 5072715 |
| library:ci-category_theory       | 10min 33s  | 69.78 MiB        | 5072769 |
| test-suite:base+async            | 10min 59s  | 0.00 B           | 5072718 |
| test-suite:edge+flambda          | 10min 13s  | 0.00 B           | 5072716 |
| library:ci-itree                 | 10min 26s  | 89.67 MiB        | 5072777 |
| library:ci-riscv_coq             | 11min 34s  | 89.53 MiB        | 5072787 |
| pkg:opam:native                  | 11min 13s  | 19.62 KiB        | 5072707 |
| plugin:ci-fiat_parsers           | 12min 59s  | 60.81 MiB        | 5072755 |
| library:ci-coq_performance_tests | 12min 17s  | 55.42 MiB        | 5072725 |
| library:ci-fiat_crypto_ocaml     | 12min 26s  | 0.00 B           | 5072792 |
| library:ci-coq_tools             | 13min 36s  | 47.03 MiB        | 5072726 |
| library:ci-mathcomp              | 14min 13s  | 217.52 MiB       | 5072776 |
| library:ci-jasmin                | 15min 46s  | 255.02 MiB       | 5072804 |
| library:ci-corn                  | 15min 15s  | 171.44 MiB       | 5072795 |
| library:ci-oddorder              | 15min 19s  | 237.85 MiB       | 5072793 |
| library:ci-verdi_raft            | 16min 15s  | 84.87 MiB        | 5072744 |
| library:ci-bedrock2              | 21min 57s  | 150.88 MiB       | 5072788 |
| doc:ci-refman                    | 22min 26s  | 24.67 MiB        | 5072811 |
| library:ci-iris                  | 22min 26s  | 108.08 MiB       | 5072774 |
| library:ci-analysis              | 23min 57s  | 265.70 MiB       | 5072800 |
| library:ci-fourcolor             | 27min 53s  | 236.46 MiB       | 5072794 |
| plugin:ci-perennial              | 31min 42s  | 125.51 MiB       | 5072760 |
| plugin:ci-metacoq                | 36min 11s  | 235.11 MiB       | 5072786 |
| library:ci-vst                   | 37min 46s  | 275.99 MiB       | 5072802 |
| library:ci-fiat_crypto           | 41min 6s   | 367.00 MiB       | 5072791 |
| library:ci-unimath               | 42min 44s  | 46.15 MiB        | 5072743 |
| library:ci-fiat_crypto_legacy    | 52min 36s  | 106.66 MiB       | 5072732 |
| library:ci-bedrock2_examples     | 52min 12s  | 186.70 MiB       | 5072789 |
| total                            | 16h 31min  | 11.40 GiB        | 1085613 |
~~~

this PR https://gitlab.inria.fr/coq/coq/-/pipelines/1085818
~~~
| name                             | duration   | artifacts size   |      id |
|----------------------------------+------------+------------------+---------|
| validate:base+32bit              | 1min 8s    | 13.70 KiB        | 5074986 |
| plugin:ci-coq_dpdgraph           | 1min 9s    | 4.14 MiB         | 5075017 |
| plugin:ci-aac_tactics            | 1min 10s   | 3.01 MiB         | 5075011 |
| plugin:ci-coinduction            | 1min 13s   | 990.10 KiB       | 5075016 |
| library:ci-paco                  | 1min 14s   | 8.11 MiB         | 5075004 |
| plugin:ci-lean_importer          | 1min 17s   | 15.99 MiB        | 5075022 |
| plugin:ci-paramcoq               | 1min 17s   | 3.96 MiB         | 5075025 |
| plugin:ci-coqhammer              | 1min 23s   | 6.65 MiB         | 5075018 |
| library:ci-simple_io             | 1min 24s   | 1.09 MiB         | 5075044 |
| plugin:ci-equations              | 2min 34s   | 9.76 MiB         | 5075020 |
| plugin:ci-bignums                | 2min 43s   | 7.51 MiB         | 5075015 |
| library:ci-menhir                | 2min 43s   | 56.74 MiB        | 5075001 |
…